Adding Parser.peek_explicit/implicit_tag()
The idea behind this change is such that when encoding an enum (via
CHOICE) using tagging, you can determine which variant to read by
peeking at the tag value without advancing the parser.

`peek_explicit_tag()` and `peek_implicit_tag()` return a tuple like

```
(<raw-tag>, <decoded-tag>)
```

such that `raw-tag` is the tag as it is read directly from the input
bytes, and `decoded-tag` is the tag converted into the same format that
would be passed to `Writer.write_explicit_element()` or
`Writer.write_implicit_element()`.

`peek_implicit_tag()` requires knowing the type you wish to decode
up-front in order to retrieve an accurate `decoded-tag` value, as
implicit tagging embeds the type within the tag.